% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMET.R
\name{printMET}
\alias{printMET}
\title{Print summary overall results from a Multi-Environment Trial analysis in a workbook}
\usage{
printMET(
  .asreml.obj,
  .fa.asreml.obj = NULL,
  .aii.pedicure = NULL,
  .print = FALSE,
  .output = "MET_output.xlsx"
)
}
\arguments{
\item{.asreml.obj}{An asreml object with the linear mixed model fitted for the analysis.}

\item{.aii.pedicure}{Set as NULL by default. If molecular marker data was used to compute the kinship
matrix during the analysis, the mean value of the diagonal elements of the kinship matrix prior to
the scaling must be introduced in this argument. The mean value of the diagonal elements of the kinship
matrix is used to compute the additive genetic variance.}

\item{.print}{If you wish to print the outputs in an excel file, please set as TRUE. Set as FALSE by default.}

\item{.output}{An *.xlsx file to write the outputs. The default is "MET_output.xlsx".}

\item{.fa.asreml.objectect}{A fa.asreml object from ASExtras4 package. Set as NULL by default. This
object is necesary only if a FA model was fitted}
}
\value{
A list containing

- $Trials_Summary a summary of the trials

- $Results_Summary a summary of the additive, non-additive and total genetic variance, error variance

- $Additive_Cmat and $Total_Cmat the additive and total genetic correlation matrices, if genetic correlation was accounted for

- $Additive_Heatmap and $Total_Heatmap the heatmaps for additive and total genetic correlation matrices, if genetic correlation was accounted for

- $Concurrence a genotype concurrence by trial table

- $GxE_tab a genotype by trial table

- $FA_sum the percentage of variance accounted for each trial by each
factor and in total if a FA model was fitted

A *.xlsx file comprised of two spreadsheet containing the information from the list
}
\description{
Prints a summary and overall results from a Multi-Environment Trial
analysis in a *.xlsx file using openxlsx. The required inputs depend on the model that was fitted and the
type of genetic information used, if any. The workbook is comprised of two worksheets. The first one contains
basic stats by trial, genotype concurrence by trial and a genotype by trial table. The second one contains
a model summary. The function also returns a list containing all the mentioned summaries.
}
