% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printST.R
\name{printST}
\alias{printST}
\title{Print summary overall results from a Single Trial analysis in a workbook}
\usage{
printST(asreml.obj, .output = "ST_output.xlsx")
}
\arguments{
\item{asreml.obj}{An asreml object with the linear mixed model fitted for the analysis.}
}
\value{
A *.xlsx file comprised of two spreadsheet containing the information from the list
}
\description{
Prints a summary and overall results from a Single Trial
analysis in a *.xlsx file using openxlsx. The workbook is comprised of two worksheets.
The first one contains basic trial stats and summaries. The second one contains
a model summary.
This function considers the following column names: Trial, Genotype, Replicate, Row, Column. If
any of these columns do not posses these respective names, please be sure of renaming them.
}
