% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printStats.R
\name{printStats}
\alias{printStats}
\title{Summary stats of recorded traits}
\usage{
printStats(.tab, .output = "SumStats_output.xlsx")
}
\arguments{
\item{.tab}{A data frame object containing the table with the summary stats of the traits of interest
for each genotype.}

\item{.output}{An *.xlsx file to write the outputs. The default is "SumStats_output.xlsx".}
}
\description{
Provides a data frame with minimum, median and maximum values from traits of
interest and the number of trials that the trait was measured for each genotype, together with pedigree
information, if any. The output from this function can be merged with predicted means from other analysed
traits. Once the table of summary statistics is complete, with or without additional predicted means,
everything should be ready to use printSumStats, the function to print this table in a nice way in an
*.xlsx file.
}
