% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTab.R
\name{printTab}
\alias{printTab}
\title{Print a table in an worksheet}
\usage{
printTab(
  .tab,
  .position = c(1, 1),
  .output = "coolTab_output.xlsx",
  .sheet = "Sheet",
  .title = NULL
)
}
\arguments{
\item{.tab}{The table to be printed}

\item{.position}{The position in the spreadsheet where the table will be printed. Set by default as 1,1.}

\item{.output}{The name of the output file. Set by default as "coolTab_output.xlsx".}

\item{.sheet}{The name of the sheet where the table will be printed. Set by default as "Sheet".}

\item{.title}{The table title. Set by default as "This is my cool tab".}
}
\description{
Prints a table in a *.xlsx file with a nice format.
}
