% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumGeno.R
\name{sumGeno}
\alias{sumGeno}
\title{Summary of genotype by trial}
\usage{
sumGeno(.df, .geno = "Genotype", .trial = "Trial", .cbf = FALSE)
}
\arguments{
\item{.df}{A dataframe object containing the data to be analysed}

\item{.geno}{The name of the genotype factor in `.df`. Set as "Genotype" by default}

\item{.trial}{The name of the trial factor in `.df`. Set as "Trial" by default}

\item{.cbf}{Set as FALSE by default. If TRUE, the function also returns a heatmap of genotype
concurrence by trial with colorblind friendly palette.}
}
\value{
A list containing

- A list with the genotype concurrence by trial table, a heatmap of this table and, if
specified, a heatmap with colorblind friendly palette;

- A data frame with the table of genotypes by trial
}
\description{
Provides information about the genotype concurrences by trial
and the genotype by trial.
}
