% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumStats.R
\name{sumStats}
\alias{sumStats}
\title{Summary stats of recorded traits}
\usage{
sumStats(.df, .trait, .geno = "Genotype", .trial = "Trial", .pedigree = NULL)
}
\arguments{
\item{.df}{A dataframe object containing the data to be analysed.}

\item{.trait}{A vector containing the index or names from the traits that you are interested in}

\item{.geno}{The name of the genotype factor in `.df`. Set as "Genotype" by default}

\item{.trial}{The name of the trial factor in `.df`. Set as "Trial" by default
summarising. Traits must be set as numeric. The function provides the mean, standard deviation and
missing values.}

\item{.pedigree}{A data frame containing the pedigree information from each genotype. Set as NULL by default.}
}
\description{
Provides a data frame with minimum, median and maximum values from traits of
interest and the number of trials that the trait was measured for each genotype, together with pedigree
information, if any. The output from this function can be merged with predicted means from other analysed
traits. Once the table of summary statistics is complete, with or without additional predicted means,
everything should be ready to use printSumStats, the function to print this table in a nice way in an
*.xlsx file.
}
