% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTrait.R
\name{sumTrait}
\alias{sumTrait}
\title{Explore the observed response of a specific trait}
\usage{
sumTrait(.df, .trait = NULL, .trial = NULL)
}
\arguments{
\item{.df}{A dataframe object containing the data to be analysed.}

\item{.trait}{The names of the trait that you are interested in. Trait must be set as numeric in the
data set.}

\item{.trial}{Set as `NULL` by default. If multiple trials are considered, please enter the name of
the trial factor.}
}
\value{
If a single trial was considered, returns a list comprised of

- a ggplot object of the observed response in the field

- a ggplot object with a histogram of the trait
}
\description{
Aims to explore the observed response of a specific trait. This function considers
simple breeding trial designs, namely, rectangular arrays with replicate blocks
and genotypes randomised within replicates. Structural terms must be named as "Replicate", "Row"
and "Column", and the genotype or variety factor as "Genotype". This function
provides summary plots of the trait of interest.
}
