% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTrial.R
\name{sumTrial}
\alias{sumTrial}
\title{Summary of data set}
\usage{
sumTrial(.df, .trait = NULL, .trial = "Trial")
}
\arguments{
\item{.df}{A dataframe object containing the data to be analysed.}

\item{.trait}{A vector containing the index or names from the traits that you are interested in
summarising. Traits must be set as numeric. The function provides the mean, standard deviation and
missing values.}

\item{.trial}{Set as "Trial" by default. If multiple trials are considered, please enter the name of the trial factor.}
}
\value{
Returns a list comprised of

- a dataframe with a summary of trial structure and traits selected

- if multiple trials, a list with a ggplot object per trial. If a single trial, a ggplot object
}
\description{
Aims to summarise trials from a data set. This function considers
simple breeding trial designs, namely, rectangular arrays with replicate blocks and genotypes randomised
within replicates. Structural terms must be named as "Replicate", "Row" and "Column", and the genotype
or variety factor as "Genotype". This function also provides plots of the trial layout
coloured by genotype levels.
}
