% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useFAST.R
\name{useFAST}
\alias{useFAST}
\title{Use Factor Analytic Selection Tools}
\usage{
useFAST(
  .asreml.obj,
  .fa.asreml.obj = NULL,
  .output = "FAST_output.xlsx",
  .sheet,
  .aii.pedicure = NULL
)
}
\arguments{
\item{.asreml.obj}{An asreml object with the linear mixed model fitted for the analysis.}

\item{.output}{An *.xlsx file to write the outputs. The default is "FAST_output.xlsx".}

\item{.sheet}{The name of the sheet where the table will be printed. Set by default as "FAST-additive" if
.nonadd = FALSE and as FAST-total if .nonadd = TRUE.}

\item{.aii.pedicure}{Set as NULL by default. If molecular marker data was used to compute the kinship
matrix during the analysis, the mean value of the diagonal elements of the kinship matrix prior to
the scaling must be introduced in this argument. The mean value of the diagonal elements of the kinship
matrix is used to compute the additive genetic variance.}

\item{.fa.asreml.object}{A fa.asreml object from ASExtras4 package. Set as NULL by default. This
object is necesary only if a FA model was fitted}
}
\value{
A list containing

- $OP the Overall Performance measure

- $RMSD the Root Mean Square Deviation measure

- $R_Fi Responsiveness for factor i. There will be as much responsiveness objects within the list as
factors -1 there are.

- $adjBLUP the adjusted BLUPs.

A spreadsheet containing the information from the list. The spreadsheet can be printed in a new or
existing *.xlsx file.
}
\description{
Prints the results from the use of Factor Analytic Selection Tools, FAST, in a *.xlsx file
using openxlsx. The function also returns a list containing the results.
}
